Update patch set 6
Patch Set 6:

Danny, re-requesting the permission may not actually do anything if the user has previously checked the "don't ask me again" box. In that case,the system settings app panel is the _only_ way to grant the permission, and requestPermission will return false without interacting with the user until that is done.

Patch-set: 6